Apply insert rules on Instances when fishing for metadata (#985)
An Instance's url may be set by a rule from a RuleSet. Apply insert
rules on the found instance so that these rules can be found in order to
correctly determine the url.

It is not necessary to do this with other FSH types because Instances
are exported after all other FSH types. So, the correct url can be found
on the exported resource in the Package.