Do not be fooled by slices with the same name as the sliced element
When finding an element by path, it is permitted to use the name of a
choice element slice (such as valueString) directly in a path. This is
achieved by considering elements to match when the element id contains
the appropriate path part and slice name (value[x]:valueString, for
example). However, this should only be supported for choice elements, so
do not consider an element to be a match when the name of the slice and
the name of the element are the same.