Don't clone original ElementDefinition's StructureDefinition (#1013)
* Don't clone original ElementDefinition's StructureDefinition

When cloning an ElementDefinition, save and remove the referenced
StructureDefinitions on the element and the element's original version.
Then, after cloning the other properties, restore those references.

* Remove structDef when capturing original

The _original is used to calculate the differential. This means that
only properties present on the JSON representation of the
ElementDefinition are used. Remove structDef from the _original in order
to simplify cloning.