Combine named slices and soft indexing
When assigning values on an Instance, it may be useful to use a
combination of soft indices and named slices. When this happens, first
create all the slices that will need to exist, both named and unnamed.
Then, proceed to apply values from the InstanceOf StructureDefinition
and any AssignmentRules on the Instance. Prior to this change, the
ordering of operations would frequently result in elements being merged
together in undesired ways. This was due to both the way soft indices
were resolved, as well as the assignment of implied properties when
exporting an Instance. Soft index resolution now accounts for named
slices when determining the indices of unnamed slices. Implied
properties are assigned after the slices are created in the correct
order, which means that they are no longer prone to forcing implied
properties on those slices to appear as the earliest entries in the
list. The test added to InstanceExporter demonstrates the difference in
behavior.