Avoid call stack overflow when assigning large base64 values
When assigning a string to a base64Binary-type element, the value is
checked to make sure it is meets the requirements of that type. The
value must be composed of valid base64 chunks. Checking this one chunk
at a time avoids the call stack overflow that may occur when checking
the entire value all at once.