Try to find required elements without slice names
When assigning values on a sliced element on an Instance, it is possible
to refer to slices by name or by numeric index. If a numeric index is
used in the rule on the Instance, but the slice being modified is a
named slice on the InstanceOf definition, the slice name will not be
tracked correctly on child elements on that Instance. If this is found,
warn the user that it is preferable to use slice names in rules when
possible, as it will make the FSH easier to understand. Additionally,
this will help ensure consistent behavior when slice ordering rules are
overhauled such that numeric-indexed slices no longer overwrite named
slices.