Handle choice paths on Instance assignment rules
When assigning values to choice elements on an Instance, type-specific
elements should be used in the rules. If the choice element is
constrained to a single type, it is possible to change the path to refer
to the type-specific element. Warn the user that the path is being
changed and that it is preferable to use the type-specific element name
in the rule. If the choice element is constrained to multiple types, it
is not possible to change the path. Show an error indicating that the
value will not be assigned. These warnings and errors may apply
regardless of where the choice element appears in the rule path.

See also  http://hl7.org/fhir/R4/formats.html#choice