Parse rules to add contentReference elements
When adding an element to a Logical or Reference, it may have a
contentReference defined. Such elements do not have a defined set of
types. Therefore, the syntax to add these elements is different than the
syntax for adding elements with types. Regardless of whether the added
element has types or a contentReference, create an AddElementRule with
the relevant information.

The grammar for adding a contentReference element is currently
experimental and may be revised.