Add old path to warning message
When a choice element path with one type gets updated, include the old
path in the message shown to the user.

When trying to update a path with choice elements, immediately return
the input path if there are no [x] present. This is faster than going
through the whole process of disassembling and reassembling the path
into the same path.