Add content reference elements on Logical and Resource definitions (#1049)
* Parse rules to add contentReference elements

When adding an element to a Logical or Reference, it may have a
contentReference defined. Such elements do not have a defined set of
types. Therefore, the syntax to add these elements is different than the
syntax for adding elements with types. Regardless of whether the added
element has types or a contentReference, create an AddElementRule with
the relevant information.

The grammar for adding a contentReference element is currently
experimental and may be revised.

* Export elements with a contentReference

When an AddElementRule has no types, this is because it has a
contentReference instead. When applying the rule, set the
contentReference on the newly created element.

* Factor out shared fields when parsing rules that add elements

* Same token order for typed and contentReference element rules

Both kinds of rules for adding elements now start with path,
cardinality, and flags.

* Token order is same in both types of add element rules

When adding a contentReference element, the rule starts with path,
cardinality, and flags, just like a typed element.

Since the contentReference is a URL, this is a reasonable place to use
an alias. Parse the value as alias-aware when creating the rule.

* Use realistic contentReferences in tests