Update antlr4 to 4.9.2 (#933)
* Update antlr4 to 4.9.2

The updated antlr4 library generates classes with a different export
style. Update the imports of those classes to use the default import
rather than a destructured import.

The directory structure of the antlr4 package adds two new directories.
Add package path mapping to the TypeScript compiler configuration and
the Jest configuration.

* Update antlr4 to 4.9.3

Update year in test fixtures, since they haven't been run since last
year.

* Handle antlr path resolution at runtime

Setting paths in the tsconfig is good for helping tsc find modules, but
doesn't help node out at runtime. Add the tsconfig-paths module so that
the modules will resolve. This works, but it adds a new library, so I'm
upset about it.

* Resolve antlr4 submodules relative to app

* Fix path registration in index.ts

When using SUSHI as a module, the entry point is index, not app. So,
both files need to have the antlr4 path registration applied.