Register plugins from configuration
A new configuration attribute is added to allow users to register
plugins. Plugins can be managed directly on the filesystem or installed
from npm. For npm plugins, a version must be included. Plugins should
export an initialize function, which will be called when the plugin is
loaded. The initialize function is passed a function that allows hooks
to be registered. Additionally, any parameters provided in the
configuration will also be passed to the initialize function.

The available hooks at which functions may be registered are:
- beforeFillTank
- beforeLoadCustomResources
- afterLoadCustomResources
- afterExportFHIR

The code to call registered functions at these points is not yet
implemented.