Register plugins from configuration (#1065)
* Register plugins from configuration

A new configuration attribute is added to allow users to register
plugins. Plugins can be managed directly on the filesystem or installed
from npm. For npm plugins, a version must be included. Plugins should
export an initialize function, which will be called when the plugin is
loaded. The initialize function is passed a function that allows hooks
to be registered. Additionally, any parameters provided in the
configuration will also be passed to the initialize function.

The available hooks at which functions may be registered are:
- beforeFillTank
- beforeLoadCustomResources
- afterLoadCustomResources
- afterExportFHIR

The code to call registered functions at these points is not yet
implemented.

* Make plugin config types match

* Improve error message when initialize function does not exist

When a plugin module is imported successfully, but does not have an
initalize function, throw MissingInitializeFunctionError.

Refactor processPluginConfiguration to have fewer places where
load/install are called.

Only attempt to process plugin configuration if it is not nullish.

* Allow simpler variations on plugin configuration

Plugin configuration now uses the name of the plugin as the key, rather
than using a "name" attribute. See example-config.yaml for examples of
permitted configuration.

Version is now specified separately from the name, so logic to parse out
the version from the name is removed.

* Handle warnings from npm during plugin installation

When an error occurs during npm install, the execFile promise is
rejected. A warning during npm install writes to stderr, but the promise
resolves. Any other messages from npm install on stderr are logged as
info messages.

Changed plugin directory to "sushi-plugins" at project root.

* Plugin initialize function may be asynchronous

* Do not save dependency info when installing from npm

* Provide info message before installation from npm