Avoid assigning duplicate values on sliced elements (#1071)
* Add required slices first when setting implied properties

When a sliced element has required slices, its minimum cardinality will
be at least the sum of the slice minimums. Add properties of named
slices before adding elements based on the minimum cardinality.

* Simplify PathNode type

* Remove unused import

* Create additional array elements when necessary during assignment

When an assigned value is an object that contains array properties,
each element of that array should be present on the instance. If there
is a matching element already, the instance does not need to change. If
the instance array contains an element that has some but not all of the
assigned element's values, merge the new values in. Otherwise, add a new
element to the instance's array.

* Additional test for array of primitive values