Check complex type values when validating assignments
When validating the a complex type value used in an AssignmentRule, the
value may be an InstanceDefinition. Do not assign the value if it would
conflict with existing values within the complex type. Assignment of
complex values is all-or-nothing: if a conflict exists on part of the
complex value, no parts of the complex value are assigned. An error is
reported on the line where the Instance is assigned, and the error
message refers to the specific value within the Instance that has a
conflict.