Instance Exporter uses configured slice name enforcement
When resolving soft indexing, a boolean parameter can be passed to
indicate whether "strict soft indexing" should be used. The value of
this parameter is taken from the SUSHI configuration. When this value is
false (the default), the behavior is consistent with previous versions
of SUSHI. When this value is true, soft indexing is resolved by
enforcing the practice of referring to slices by name.