Allow cardinality changes on connected elements
When constraining cardinality on an element, if there are connected
elements that are the child of a slice, it may be possible to apply the
cardinality constraint to those elements. As long as no cardinality is
widened, applying the constraints is allowed. This may mean that only
the new min or max are applied. If the new cardinality has no overlap
with the connected element's existing cardinality, the cardinality
constraint can't be applied.