Reslices in soft indexing increment parent slices
When a path uses soft indexing to increment the value of an element with
reslices, the path part with the soft index will contain multiple
slices. The tracked index for each less-sliced version of the element
should be incremented in order to maintain the correct index when
less-sliced versions of the element are used.