Find connected elements that end with a slice
A connected element may end with a slice if the original element ends
with a slice with the same name.

When applying the "must support" flag on an element whose path ends with
a slice, the flag should be applied to connected elements only if the
connected element ends with a slice with the same name.