Allow cardinality constraints to be applied to connected elements (#1120)
* Allow cardinality changes on connected elements

When constraining cardinality on an element, if there are connected
elements that are the child of a slice, it may be possible to apply the
cardinality constraint to those elements. As long as no cardinality is
widened, applying the constraints is allowed. This may mean that only
the new min or max are applied. If the new cardinality has no overlap
with the connected element's existing cardinality, the cardinality
constraint can't be applied.

* Remove test for old behavior

* Find connected elements that end with a slice

A connected element may end with a slice if the original element ends
with a slice with the same name.

When applying the "must support" flag on an element whose path ends with
a slice, the flag should be applied to connected elements only if the
connected element ends with a slice with the same name.

* Update test/export/StructureDefinitionExporter.test.ts