Add comment to first line of generated artifact files
Check the contents of files that exist at the paths used by generated
artifact files. If the contents start with the comment, then the file is
likely a previous version of the generated file, and it is safe to
overwrite it with new generated contents. Otherwise, emit an error and
do not change the file.