Fix problem w/ regression extract and copy
The extract and copy approach used when getting repo sources was causing some unintended diffs.  It seems the copy is not as reliable as it should be.  Instead, we now extract the zip twice so that each copy of the repo is exactly the same.

This also updates the repos-all.txt file.