Order IG resources based on config (#1118)
* Order IG resources based on config

If all generated resources are present in the config resources section,
use the order from that section when adding resources to the IG.
Otherwise, if all generated resources are present in the config groups
section, use the order from that section when adding resources to the
IG. If neither config section contains all resources, add the resources
sorted by title or id.

* Fix Minifier Dependency and Location Type (#1113)

Changed dependency html-minifier to html-minifier-terser

* Add test for using title to sort resource

* Sort resources after all resources are added

Resources in the IG are sorted after adding generated, configured, and
predefined resources. A temporary sort key attribute is added to the
resources when adding them. This attribute is used for the default
sorting method.

* Refactor IG resource sorting

Test if a sort should be used and use it in the same function. This
avoids having to iterate and search over the set of resources additional
times.

* Sort resources by name

The sort key that was used to sort IG resources is removed. Resources
are now sorted by name. The name attribute should always exist, but if
it doesn't, use the reference key instead.

* Default resource order is case-insensitive

Co-authored-by: Alex Walley <alexgwalley@gmail.com>