Add command-line option to update dependencies
When this option is provided, SUSHI will attempt to update the
dependencies listed in the FSH project's configuration file.
Dependencies listed with a version of "current" or "dev" will not be
modified. The user is prompted before any updates are made and is shown
a list of all of the updates that will be applied. If the configuration
was not imported from a sushi-config file, this option does nothing.

If the user provides this flag and applies updates, the project's
sushi-config file will be rewritten with the updated version
information. This may also result in some formatting changes to the
configuration file, but all other file contents (including comments)
will be preserved.