Slice minimums include numeric indices
When calculating the minimum number of elements for sliced elements,
include numeric indices for parts of the path before the last part. This
helps avoid creating unnecessary sibling elements.

A side effect of this is that implicit extension slices created on
instances will no longer automatically fill in their URL on elements in
the list prior to the ones actually referred to in rules. However, both
the previous and new behaviors are equally erroneous, and indicate that
the author should revise their FSH to create valid extension elements.

Some cleanup still needed.