Add command-line option to update dependencies (#1152)
* Add command-line option to update dependencies

When this option is provided, SUSHI will attempt to update the
dependencies listed in the FSH project's configuration file.
Dependencies listed with a version of "current" or "dev" will not be
modified. The user is prompted before any updates are made and is shown
a list of all of the updates that will be applied. If the configuration
was not imported from a sushi-config file, this option does nothing.

If the user provides this flag and applies updates, the project's
sushi-config file will be rewritten with the updated version
information. This may also result in some formatting changes to the
configuration file, but all other file contents (including comments)
will be preserved.

* Add suggested review changes

Prompt uses Y/N instead of A/Q.
If configuration specifies additional properties for a dependency, keep
those properties when updating the dependency.
If a package is not found on the packages server, try the packages2
server.

* Always exit after checking for updates

Do not process any FSH when the update-dependencies flag is set.
Add some more info messages to explain various reasons why dependencies
may not be updated.
Check both valid file extensions for configuration file before exiting
when no FSH files are present.

* Fix URL for packages2 server

Messages that appear when version information is missing upgraded from
info to warning.
Clean up Y/N prompt shown when dependency updates are available.

* Add update-dependencies to README