Traverse SD elements to find implied properties
A new function for setting implied properties is added. Rather than
iterating over the set of rule paths and using those paths to find
elements, this function iterates over the elements. The rule paths are
still used for certain checks.

This implementation is not yet complete, and is in need of significant
cleanup. The new function is only used when exporting instances, so the
previous function will remain for now.