Automatically load hl7.fhir.uv.tools dependency (#1186)
* Automatically load hl7.fhir.uv.tools dependency

The current version of this package will always be loaded, even if it is
not configured as a dependency. If it is configured as a dependency, use
the configured version. When it fails to load, log a warning instead of
an error if it was not a configured dependency.

* Clarify warning when automatic dependency fails to load

Add helper function to check for the presence of automatic dependencies
in the package list.