Check concepts during export, not import
The addConcept and checkConcept functions are removed from
FshCodeSystem. ConceptRules are now added directly to the rules array
during import and insert rule application.
CodeSystemExporter checks each ConceptRule as it is being added to the
FHIR CodeSystem to avoid duplicates and ensure correct hierarchies. This
occurs after insert rules are applied.
Removed tests that checked for errors produced by checkConcept. Other
minor test adjustments to handle the change in where concepts are
checked. Added test to check that an ConceptRule on a CodeSystem can
refer to a hierarchy defined by a ConceptRule within an inserted
RuleSet.