Warn on indexed array items that likely match a slice
Detect cases where an array item in an instance matches a required (implied) slice but was not assigned using a slice name. We want to detect and warn on this because it very often results in undesirable outputs like repeated and/or incomplete array items. Before manualSliceOrdering, most of these cases just overwrote the implied slice instead of being prepended before the slice.

Fixes #1180