Find entities by id, even when their id is set by a rule (#1198)
* Get Instance id from FSH definition

When fishing for an Instance, the id of the Instance may be set by a
rule. The rule could be directly on the Instance, or it might be
contained within a RuleSet. Examine these rules to find potential
matches.

* Apply insert rules before any exports

When exporting FHIR, apply insert rules on all FSH definitions before
exporting any resources. Each exporter provides a method to apply insert
rules on all of the definitions for the resource type it exports. This
allows functions that are called during export to assume that all insert
rules have already been applied, which can help detect rules that
override metadata.

FSH entities that can be fished by id now have a get method for their id
that accounts for the presence of id-setting rules.

* minor tidying

* Add findIdCaretRule

Small helper function to find a CaretValueRule that provides an id.

* Add findIdAssignmentRule

Instances can have an id set by an AssignmentRule, so they get a
slightly different helper function.
Refactor replaceReferences function and Instance class to use this new
helper function.
HasId.validateId now looks for an id AssignmentRule when operating on an
Instance, which helps it give more accurate line numbers in error logs.