Use inserted rules when checking new concepts (#1196)
* Use inserted rules when checking new concepts

When applying insert rules on a CodeSystem, provide the expanded rules
when checking new concepts. This way, concepts added by RuleSets can
provide a hierarchy. Additionally, this allows for duplicate concepts to
be detected when they are added by RuleSets.

* Check concepts during export, not import

The addConcept and checkConcept functions are removed from
FshCodeSystem. ConceptRules are now added directly to the rules array
during import and insert rule application.
CodeSystemExporter checks each ConceptRule as it is being added to the
FHIR CodeSystem to avoid duplicates and ensure correct hierarchies. This
occurs after insert rules are applied.
Removed tests that checked for errors produced by checkConcept. Other
minor test adjustments to handle the change in where concepts are
checked. Added test to check that an ConceptRule on a CodeSystem can
refer to a hierarchy defined by a ConceptRule within an inserted
RuleSet.

* Add tests for duplicate code and wrong hierarchy

Small refactor in applyInsertRule for handling ConceptRule inserted with
a path context.

* Remove unnecessary error classes

Use a map to store existing concepts for simpler duplicate checks.

* Remove unused import