FSH may contain numbers in scientific notation
A number can be expressed in scientific notation by ending it with "E"
followed by the exponent. The "E" can be lower or upper case. The
exponent may contain a sign. A negative exponent is indicated by using a
minus sign. A positive exponent is indicated by using a plus sign or no
sign.