Warn on array items that match required slices but don't use a slice name (#1200)
* Add ASDF .tool-versions

* Remove --maxWorkers from test script since it interferes w/ some VSCode plugins

* Warn on indexed array items that likely match a slice

Detect cases where an array item in an instance matches a required (implied) slice but was not assigned using a slice name. We want to detect and warn on this because it very often results in undesirable outputs like repeated and/or incomplete array items. Before manualSliceOrdering, most of these cases just overwrote the implied slice instead of being prepended before the slice.

Fixes #1180

* Add a test to ensure paths w/ slicenames don't get slice match warnings

* Address review feedback (part 1)

- Use destructured assignment on collected values
- Remove unnecessary flatten and clarify argument types
- Properly detect exact matches w/ multiple slices involved
- Update warning message on exact matches
- Add more tests around exact matches

* Address review feedback (part 2)

- Remove matchesPattern function and use lodash isMatch directly
- Remove matchesPattern tests and add corresponding test in InstanceExporter test suite

* Add test-specific tsconfig

VS Code works best when the test folder has its own specific tsconfig.

* Truncate values > 300 characters in slice name warning message