Number using exponents may be very large
When parsing a NUMBER token that is used as the value for an
AssignmentRule or CaretValueRule, create a BigInt if the value is an
integer. Check the whole, decimal, and exponent parts of the NUMBER
token to determine if the value being expressed is an integer or not.

TypeScript compilation now targets ES2016 in order to support the
exponentiation operator.