FSH may contain numbers in scientific notation (#1203)
* FSH may contain numbers in scientific notation

A number can be expressed in scientific notation by ending it with "E"
followed by the exponent. The "E" can be lower or upper case. The
exponent may contain a sign. A negative exponent is indicated by using a
minus sign. A positive exponent is indicated by using a plus sign or no
sign.

* Number using exponents may be very large

When parsing a NUMBER token that is used as the value for an
AssignmentRule or CaretValueRule, create a BigInt if the value is an
integer. Check the whole, decimal, and exponent parts of the NUMBER
token to determine if the value being expressed is an integer or not.

TypeScript compilation now targets ES2016 in order to support the
exponentiation operator.

* farewell to Guhan