Support versions in instance meta.profile
When listing example instances in an IG, the instance's meta.profile is
used if it is available. The URL in meta.profile may contain a version.
If it does, check to make sure that the specified version matches the
version of the profile. If so, set exampleCanonical to the unversioned
URL. If the versions do not match, set exampleBoolean to true.