Add elements with numeric names
An element that is added to a Logical or Resource may have a name that
contains only numeric characters. This is allowed, but not recommended,
as such a name will not match the FHIR-recommended pattern for element
names. Show a warning indicating that the element name is not
recommended when adding these elements. If the element has children, do
not show the same warning again if the new part of the name matches the
FHIR-recommended pattern.