Use miniparser for RuleSet value substitutions
A miniparser is defined with a minimized set of the FSH grammar. It
contains only enough rules to try to parse a parameterized RuleSet and
perform substitutions. The miniparser is used by MiniFSHImporter to
transform a RuleSet with a set of parameter values.

Substitutions that take place within an insert rule's parameter list
should be bracketed.