Check ElementDefinition invariants related to element name when adding elements to Logical or Resource (#1230)
* Add elements with numeric names

An element that is added to a Logical or Resource may have a name that
contains only numeric characters. This is allowed, but not recommended,
as such a name will not match the FHIR-recommended pattern for element
names. Show a warning indicating that the element name is not
recommended when adding these elements. If the element has children, do
not show the same warning again if the new part of the name matches the
FHIR-recommended pattern.

* Check new element names against both ElementDefinition invariants

The FHIR specification for ElementDefinition has two invariants related
to element names. eld-19 defines prohibited characters and length
requirements. eld-20 defines a recommended pattern for element names. If
eld-19 is not met, log an error. If eld-19 is met but eld-20 is not met,
log a warning.

* Improve error and warning messages for new element paths

When checking the length of an choice element's name, do not count the [x]
against it.
Fixed up the prohibited characters expression: literal backslash was present twice.

---------

Co-authored-by: Mint Thompson <mint@mitre.org>