InstanceOf can't be abstract
When exporting an Instance, check if InstanceOf is abstract. If it is,
log an error and do not create the Instance.

When exporting a Profile, check if Parent is an abstract resource. If it
is, log an error and do not create the Profile. An abstract Profile is
allowed as a parent, however.

Remove unused InvalidProfileParentError.

Small updates to existing tests for InstanceExporter and
StructureDefinitionExporter so that they are no longer breaking these
new rules.