Don't create Instances if the nearest specialization ancestor is abstract (#1237)
* InstanceOf can't be abstract

When exporting an Instance, check if InstanceOf is abstract. If it is,
log an error and do not create the Instance.

When exporting a Profile, check if Parent is an abstract resource. If it
is, log an error and do not create the Profile. An abstract Profile is
allowed as a parent, however.

Remove unused InvalidProfileParentError.

Small updates to existing tests for InstanceExporter and
StructureDefinitionExporter so that they are no longer breaking these
new rules.

* Change to rules for abstract profile parent and instanceOf

A profile may have an abstract parent. That is fine.

When exporting an Instance, find its nearest ancestor that is a
specialization. If that ancestor is abstract, throw an error.