Allow assignment of Instance values in any entity
Rules that assign a value, with or without a caret path, may specify an
Instance as the value to assign. This applies to Structure Definitions,
Code Systems, Value Sets, and Instances.

An Instance may have an id that is parsed as a numeric or boolean value.
In these cases, if assignment of the numeric value fails, try to find an
Instance by fishing using that numeric or boolean value. If an Instance
is found, try to assign that Instance. The "raw value" of the number or
boolean from the rule is used when fishing. For example, if the number
150 was specified using exponent notation as "15E1", then an Instance
with id "15E1" may be found. An Instance with id "150" would not be
found.

If an Instance with the numeric or boolean id is found, but the
assignment or validation of the Instance value fails, the error message
to log depends on what kind of failure occurred. If a type mismatch
occurred, log an error message with the original type of the value, not
the Instance type. If a different type of error occurred when assigning
the Instance, log a message for that error.

When printing the FSH representation of Assignment or Caret Value rules,
use the raw numeric or boolean value.