Insert rule parameters can be enclosed in double brackets (#1216)
* Insert rule parameters can be enclosed in double brackets

When a parameter is enclosed in double brackets, commas and right
parentheses do not need to be escaped. The entire parameter value must
be enclosed for this feature to be used. Existing insert rules are not
affected.

* Handle literal backslash followed by comma, inside brackets

* Use miniparser for RuleSet value substitutions

A miniparser is defined with a minimized set of the FSH grammar. It
contains only enough rules to try to parse a parameterized RuleSet and
perform substitutions. The miniparser is used by MiniFSHImporter to
transform a RuleSet with a set of parameter values.

Substitutions that take place within an insert rule's parameter list
should be bracketed.

* Improve bracket-aware substitution

A substitution sequence may be within brackets, but not immediately
adjacent to the brackets. If the substitution is anywhere within
brackets, it needs to have the bracket escapes applied, but since the
brackets won't be part of the RegExp capture, they don't need to be
reapplied.

* Maintain indented rules in parameterized rule sets

Changes to how the main parser handles contents of parameterized rule
sets, as well as changes to the STAR token, means that when manually
setting a ParamRuleSet's contents in a test case, there needs to be a
leading EOL.

* Comment cleanup

* Add parser rules for parameterized rule set references

The new parser rules provide each parameter in a RuleSet definition or
in an insert rule as a separate token. This significantly simplifies the
processing that occurs for these parser contexts in FSHImporter. Thank
you to @cmoesel for the improved grammar.

* Support unicode in mini parser

Update dependency versions that were previously downgraded somehow.

* Use double-brackets when writing FSH for InsertRules

* Test with double bracket escapes

* tidy up regular expressions