Convert primitive to object only if assigned with a slice name
When assigning a primitive value on an Instance, if it is being assigned
as an element in an array, and it has a slice name, convert it into an
object immediately before assigning the value. In the previous
implementation, the conversion happened as soon as a slice name was
encountered anywhere in the assignment path, which meant that the
assigned value would be an object even in cases where it was not
assigned as an array element. This led to undesired behavior when a
later assignment path did not include a slice name, leading to the
primitive assignment value being assigned as if it were an object (that
is, using assignComplexValue).