Warn on unrecognized configuration properties
An unrecognized configuration property is likely to represent something
like a misspelled property, so emit a warning to bring it to the
author's attention. Similarly, a mis-indented property may end up in the
list of pages, so emit a warning when an entry in the list of pages has
a URL that doesn't have .md or .xml at the end.