Use JSON schema to detect potential mistakes
A JSON schema is defined that is approximately equivalent to the type
definition for the YAML configuration. It is less strict, since it is
not designed for full validation. Emit a warning for each property that
has schema validation errors. The message is different if the schema
validation error is due to an additional property, since that usually
represents a mistake in spelling or indentation.