Warn on unrecognized configuration properties and on configured pages with unusual names (#1265)
* Warn on unrecognized configuration properties

An unrecognized configuration property is likely to represent something
like a misspelled property, so emit a warning to bring it to the
author's attention. Similarly, a mis-indented property may end up in the
list of pages, so emit a warning when an entry in the list of pages has
a URL that doesn't have .md or .xml at the end.

* Use JSON schema to detect potential mistakes

A JSON schema is defined that is approximately equivalent to the type
definition for the YAML configuration. It is less strict, since it is
not designed for full validation. Emit a warning for each property that
has schema validation errors. The message is different if the schema
validation error is due to an additional property, since that usually
represents a mistake in spelling or indentation.

* Move schema to .json file

Provide unique message when additional properties are found on a path
that starts with "pages", since that property has patternProperties
defined.

* Make schema file an actual schema

* Add some missing properties and types to config schema