Validate extension context during export
When using the Context keyword on an Extension, validate that the
extension or element specified is something that actually exists. The
element is specified by a FSH path, but is exported as a FHIR element
id. Always use the resource id when the context is an element on a core
FHIR resource. Otherwise, use the URL followed by an element id.