Support Context keyword when defining Extensions (#1282)
* Import Extensions using Context keyword

The Context keyword is used to specify the contexts in which the
Extension should be used. Add the keyword to the grammar definition.
Store contexts with source info when importing FSH.

* Set context when exporting Extensions

Contexts specified with the Context keyword are applied to Extensions
during export. If the value is quoted, it is a "fhirpath" context. If
the value is not quoted and resolves to an Extension definition, it is
an "extension" context. Otherwise, it is an "element" context.

* Use Context keyword only once

The Context keyword is used at most once per extension. This is similar
to other metadata keywords. To provide multiple contexts, a
comma-separated list can be provided.

* Validate extension context during export

When using the Context keyword on an Extension, validate that the
extension or element specified is something that actually exists. The
element is specified by a FSH path, but is exported as a FHIR element
id. Always use the resource id when the context is an element on a core
FHIR resource. Otherwise, use the URL followed by an element id.

* Find deep contained extensions

A complex extension may contain more than one level of extension
elements. Traverse the extension by url to find deeply nested
extensions.

When an element path is given as context, it may represent a deeply
nested extension. If so, set the context as extension context. The
element counts as a deeply nested extension if every element along its
path is also an extension.

* comment cleanup

* Remove special syntax for contained extensions

Contained extensions must be specified the same as any other element by
using a FSH path.

* Keep parent context if Context keyword is not used

Only set the default context if the Context keyword is unused and there
is no existing context.

Assume that the last # present in an unquoted context is what separates
the URL from the path. This helps with URLs that contain a # character.

* Extension toFSH only writes Context keyword once

* Discard extra whitespace tokens in list of contexts

Remove unused code from Extension toFSH method.