Assume non-primitive if an element won't unfold
An element with a content reference may fail to unfold if there are
problems in the element's StructureDefinition or the referenced
element's StructureDefinition, or if the content reference does not
resolve to an element. In these cases, assume the element is not a
primitive for the purpose of validating an instance.