Set characteristics when exporting logical models
Logical model definitions with characteristics will include those
characteristics in the exported StructureDefinition's extension element.
The characteristics are checked against the type-characteristics-code
CodeSystem, and a warning is emitted if the characteristic is not
present in the CodeSystem. An exception is made for the "can-be-target"
characteristic, which was unintentionally left out of the CodeSystem. If
this CodeSystem is not available, a warning is emitted to let the author
know that their characteristics are not being checked.