Slicing rules should not be on slices (#1291)
* Slicing rules should not be on slices

You can add reslices of a slice without having to define slicing rules
on the slice. If you want more discriminators, you set those on the base
element. Show a warning if the user adds slicing rules to an element
that has a slice name.

* Narrow type of ValidationError severity

Severity is a string that can be used when calling logger.log, which
makes working with those errors more convenient.

When sliceIt is called on a slice, the slicing will still be added. This
puts it in line with cases where the slicing is added via caret rules.

* Update src/fhirtypes/ElementDefinition.ts