Assignment on primitive satisfies cardinality of value element (#1284)
* Assignment on primitive satisfies cardinality of value element

All elements with primitive types have a child element named "value".
When this child element is required, an assignment to the primitive
satisfies this cardinality requirement.

* Handle required values on primitive arrays

There is an existing problem regarding primitive arrays that is outside
the scope of this change. Test cases are added and skipped that show
these cases.

* Assume non-primitive if an element won't unfold

An element with a content reference may fail to unfold if there are
problems in the element's StructureDefinition or the referenced
element's StructureDefinition, or if the content reference does not
resolve to an element. In these cases, assume the element is not a
primitive for the purpose of validating an instance.

* Avoid unfolding when checking if an element is a primitive

Unfolding the full content reference doesn't provide any benefit, since
the type is all that is needed to determine whether an element is a
primitive or not.