Only validate characteristics during export
Characteristic validation is removed from the importer. If an invalid
characteristic code is specified, a warning is logged during export. The
export warning lists the supported characteristic codes.

Remove hard-coded list of type characteristic codes. Use code system
definition to get list of type characteristic codes. Manually add
"can-be-target" to the list of supported codes if it is not present in
the code system. The list of supported codes is created in the
StructureDefinitionExporter constructor.