Check referenced logicals for target validity
A logical model should specify the can-be-target characteristic for it
to be used as the target of a Reference. When constraining Reference
types, check to make sure that logical models that can be referenced
specify this characteristic. If there are any that do not specify this
characteristic, apply the type constraint and log a warning. When
assigning a specific Reference, check its type. If its type is a logical
model, and it does not specify the can-be-target characteristic, perform
the assignment and log a warning.