Reduce amount of fishing when replacing references
When a referenced item is an Instance, we can usually get the metadata
for its type pretty early. Keep a reference to it for use in the logical
target check.

Try to set canBeTarget when fishing for metadata in FHIRDefinitions and
Package objects. If fishing for metadata for a logical model in a
Package, the extensions for indicating a valid reference target may not
exist if the export of that logical model is still in progress. In that
case, return no metadata to force fishing in the FSHTank as a fallback.

Move type characteristics and logical target URLs to fhirdefs to
simplify import structure.